move functions to OO style
