+++ crates/notion-core/src/toolchain/mod.rs
@@ -3,7 +3,6 @@ use std::io::Write;
 
 use readext::ReadExt;
 use semver::Version;
-use toml;
 
 use distro::node::NodeVersion;
 use fs::touch;
@@ -18,20 +17,12 @@ pub struct Toolchain {
     platform: Option<Image>,
 }
 
-fn platform_from_toml_str(src: String) -> Fallible<serial::Platform> {
-    toml::from_str(&src).unknown()
-}
-
-fn platform_to_toml_str(platform: &Image) -> Fallible<String> {
-    toml::to_string_pretty(&platform.to_serial()).unknown()
-}
-
 impl Toolchain {
     pub fn current() -> Fallible<Toolchain> {
         let path = user_platform_file()?;
         let src = touch(&path)?.read_into_string().unknown()?;
         Ok(Toolchain {
-            platform: platform_from_toml_str(src)?.into_image()?,
+            platform: serial::Platform::from_toml(src)?.into_image()?,
         })
     }
 
@@ -90,7 +81,7 @@ impl Toolchain {
         let mut file = File::create(&path).unknown()?;
         match &self.platform {
             &Some(ref platform) => {
-                let src = platform_to_toml_str(platform)?;
+                let src = platform.to_serial().to_toml()?;
                 file.write_all(src.as_bytes()).unknown()?;
             }
             &None => {
@@ -101,48 +92,3 @@ impl Toolchain {
     }
 }
 
-#[cfg(test)]
-pub mod tests {
-
-    use self::serial;
-    use super::*;
-    use distro;
-    use image;
-    use semver;
-
-    const BASIC_TOML_STR: &'static str = r#"yarn = '1.2.3'
-
-[node]
-runtime = '4.5.6'
-npm = '7.8.9'
-"#;
-
-    #[test]
-    fn test_platform_from_toml_str() {
-        let toml_str = BASIC_TOML_STR.to_string();
-        let platform = platform_from_toml_str(toml_str).expect("could not parse TOML string");
-        let expected_platform = serial::Platform {
-            yarn: Some("1.2.3".to_string()),
-            node: Some(serial::NodeVersion {
-                runtime: "4.5.6".to_string(),
-                npm: "7.8.9".to_string(),
-            }),
-        };
-        assert_eq!(platform, expected_platform);
-    }
-
-    #[test]
-    fn test_platform_to_toml_str() {
-        let platform = image::Image {
-            yarn: Some(semver::Version::parse("1.2.3").expect("could not parse semver version")),
-            node: distro::node::NodeVersion {
-                runtime: semver::Version::parse("4.5.6").expect("could not parse semver version"),
-                npm: semver::Version::parse("7.8.9").expect("could not parse semver version"),
-            },
-        };
-        let toml_str =
-            platform_to_toml_str(&platform).expect("could not serialize platform to TOML");
-        let expected_toml_str = BASIC_TOML_STR.to_string();
-        assert_eq!(toml_str, expected_toml_str);
-    }
-}
